requirements.txt
numpy
flask
tensorflow
matplotlib
base64
Pillow
keras
gunicorn